Change int32_t to int64_t for window info

+++ packages/video_player_avplay/tizen/src/media_player.cc
@@ -430,10 +430,12 @@ bool MediaPlayer::SetDisplay(const flutter::EncodableMap *window_geometry) {
     return false;
   }
 
-  int x = flutter_common::GetValue(window_geometry, "x", 0);
-  int y = flutter_common::GetValue(window_geometry, "y", 0);
-  int width = flutter_common::GetValue(window_geometry, "width", 0);
-  int height = flutter_common::GetValue(window_geometry, "height", 0);
+  int64_t x = flutter_common::GetValue(window_geometry, "x", (int64_t)0);
+  int64_t y = flutter_common::GetValue(window_geometry, "y", (int64_t)0);
+  int64_t width =
+      flutter_common::GetValue(window_geometry, "width", (int64_t)0);
+  int64_t height =
+      flutter_common::GetValue(window_geometry, "height", (int64_t)0);
 
   if (width <= 0 || height <= 0) {
     LOG_ERROR("[MediaPlayer] Invalid window geometry: width=%d, height=%d.",
